move pull logic to lib
